toolchain-shar-template.sh: do not offer sudo as an option
Issue: LIN1018-1376

We do not want to offer our users the ability to automatically run sudo.

Using sudo from within a script can be considered a potential security issue by
some customers.

This also ensures that the person doing the installation understands the
ramifications of where they are installing the code.

WR SDK extensions may not work properly if owned by root as well.

Original author is Jeffrey C Honig.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20141114

Signed-off-by: Kai Kang <kai.kang@windriver.com>

Rebased on 20241030.

Signed-off-by: Kai Kang <kai.kang@windriver.com>