base.bbclass: add WRL_PRESERVE_BUILDTIME switch
Issue: LINCD-11604

Add WRL_PRESERVE_BUILDTIME switch to preserve buildtime.
Currently SOURCE_DATE_EPOCH usually uses the date of commit or
the date of source code file[1]. This is needed for reproducible
build. But in such case, it's also not possible to preserve
the build time of packages. So we add a switch here to set it
to the actual build time.

This is a WRLinux specific change, as Yocto does not want to
maintain two paths.

[1] https://reproducible-builds.org/docs/source-date-epoch/

(LOCAL REV: NOT UPSTREAM) -- WRLinux Specific Patch

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>