qemu.bbclass: move to classes/ directory so that INHERIT can use it
Issue: LINUXEXEC-33696

The functions in this bbclass make running binaries via qemu user mode
easier. These functions are generally useful, regardless of how this
bbclass is inherited (via "inherit" or "INHERIT"). So move it to classes/
directory so that these functions could be used by both individual recipes
and other bbclass that are intended to be used via INHERIT.

[
  Note that this is a WRLinux specific patch because oe-core upstream wants
  to refactor this bbclass into meta/lib/oe. This will affect a few other
  classes and recipes. It's unlikely such patch will be backported to walnascar
  branch. In case it is somehow backported, this local patch could be dropped.
]

(LOCAL REV: NOT UPSTREAM) -- WRLinux Specific Patch

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>