systemd: check SERIAL_CONSOLES for serial-getty-generator PACKAGECONFIG
Issue: LINCD-22981

If SERIAL_CONSOLES is empty, then systemd-serialgetty recipe does basically
nothing, and making systemd RDEPENDS on it does not help.

So a better default behavior is that if SERIAL_CONSOLES is empty, then
we put serial-getty-generator in PACKAGECONFIG which let systemd to
generate corresponding service instances.

(LOCAL REV: NOT UPSTREAM) -- sent to oe-core on 20250313

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>