qemu: fix CVE-2024-8354
Issue: LIN1025-213

A flaw was found in QEMU. An assertion failure was present in
the usb_ep_get() function in hw/net/core.c when trying to get
the USB endpoint from a USB device. This flaw may allow a
malicious unprivileged guest user to crash the QEMU process
on the host and cause a denial of service condition.

Reference:
https://nvd.nist.gov/vuln/detail/CVE-2024-8354

Upstream patch:
https://gitlab.com/qemu-project/qemu/-/commit/746269eaae16423572ae7c0dfeb66140fa882149

(LOCAL REV: NOT UPSTREAM) -- Not applicable, upstream out of maintenance

Signed-off-by: Yogita Urade <yogita.urade@windriver.com>