python3-urllib3: fix CVE-2025-66418 CVE-2025-66471
Issue: LIN1025-6279 LIN1025-6280

CVE-2025-66418: urllib3 is a user-friendly HTTP client library for
Python. Starting in version 1.24 and prior to 2.6.0, the number of links
in the decompression chain was unbounded allowing a malicious server to
insert a virtually unlimited number of compression steps leading to high
CPU usage and massive memory allocation for the decompressed data. This
vulnerability is fixed in 2.6.0.

CVE-2025-66471: urllib3 is a user-friendly HTTP client library for
Python. Starting in version 1.0 and prior to 2.6.0, the Streaming API
improperly handles highly compressed data. urllib3's streaming API is
designed for the efficient handling of large HTTP responses by reading
the content in chunks, rather than loading the entire response body into
memory at once. When streaming a compressed response, urllib3 can
perform decoding or decompression based on the HTTP Content-Encoding
header (e.g., gzip, deflate, br, or zstd). The library must read
compressed data from the network and decompress it until the requested
chunk size is met. Any resulting decompressed data that exceeds the
requested amount is held in an internal buffer for the next read
operation. The decompression logic could cause urllib3 to fully decode a
small amount of highly compressed data in a single operation. This can
result in excessive resource consumption (high CPU usage and massive
memory allocation for the decompressed data.

References:
https://nvd.nist.gov/vuln/detail/CVE-2025-66418
https://nvd.nist.gov/vuln/detail/CVE-2025-66471

(LOCAL REV: NOT UPSTREAM) -- Not applicable, upstream out of maintenance

Signed-off-by: Jiaying Song <jiaying.song.cn@windriver.com>